fix mongodb tests
