--- src/test/java/it/gov/pagopa/rtd/ms/enrolledpaymentinstrument/configs/MongoDbTest.java
@@ -1,16 +1,0 @@
-package it.gov.pagopa.rtd.ms.enrolledpaymentinstrument.configs;
-
-import org.springframework.boot.test.autoconfigure.data.mongo.DataMongoTest;
-import org.springframework.context.annotation.Import;
-import org.springframework.test.annotation.DirtiesContext;
-
-import java.lang.annotation.*;
-
-@Target(ElementType.TYPE)
-@Retention(RetentionPolicy.RUNTIME)
-@DataMongoTest
-@Inherited
-@Import({MongodbIntegrationTestConfiguration.class})
-@DirtiesContext(classMode = DirtiesContext.ClassMode.AFTER_CLASS)
-public @interface MongoDbTest {
-}
