+++ src/test/java/it/gov/pagopa/rtd/ms/enrolledpaymentinstrument/infrastructure/persistence/TkmPaymentInstrumentServiceIntegrationTest.java
@@ -5,7 +5,6 @@ import it.gov.pagopa.rtd.ms.enrolledpaymentinstrument.application.TkmPaymentInst
 import it.gov.pagopa.rtd.ms.enrolledpaymentinstrument.application.command.TkmRevokeCommand;
 import it.gov.pagopa.rtd.ms.enrolledpaymentinstrument.application.command.TkmUpdateCommand;
 import it.gov.pagopa.rtd.ms.enrolledpaymentinstrument.common.DomainEventPublisher;
-import it.gov.pagopa.rtd.ms.enrolledpaymentinstrument.configs.MongoDbTest;
 import it.gov.pagopa.rtd.ms.enrolledpaymentinstrument.configurations.RepositoryConfiguration;
 import it.gov.pagopa.rtd.ms.enrolledpaymentinstrument.domain.entities.EnrolledPaymentInstrument;
 import it.gov.pagopa.rtd.ms.enrolledpaymentinstrument.domain.entities.HashPan;
@@ -18,16 +17,22 @@ import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.runner.RunWith;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.test.autoconfigure.data.mongo.DataMongoTest;
 import org.springframework.boot.test.mock.mockito.MockBean;
 import org.springframework.context.annotation.Import;
 import org.springframework.data.domain.Sort;
 import org.springframework.data.mongodb.core.MongoTemplate;
 import org.springframework.data.mongodb.core.index.Index;
+import org.springframework.test.context.DynamicPropertyRegistry;
+import org.springframework.test.context.DynamicPropertySource;
 import org.springframework.test.context.junit4.SpringRunner;
 
 import java.util.List;
 import java.util.Set;
 import java.util.stream.Collectors;
+import org.testcontainers.containers.MongoDBContainer;
+import org.testcontainers.junit.jupiter.Container;
+import org.testcontainers.junit.jupiter.Testcontainers;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertTrue;
@@ -35,7 +40,8 @@ import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.Mockito.doReturn;
 
 @RunWith(SpringRunner.class)
-@MongoDbTest
+@DataMongoTest
+@Testcontainers
 @Import(RepositoryConfiguration.class)
 class TkmPaymentInstrumentServiceIntegrationTest {
 
@@ -105,5 +111,13 @@ class TkmPaymentInstrumentServiceIntegrationTest {
     paymentInstrumentService.handle(new TkmRevokeCommand("taxCode", hashPan.getValue(), ""));
 
     assertTrue(repository.findByHashPan(hashPan.getValue()).get().isRevoked());
+  }
+
+  @Container
+  public static final MongoDBContainer mongoContainer = new MongoDBContainer("mongo:4.4.4");
+
+  @DynamicPropertySource
+  static void setProperties(DynamicPropertyRegistry registry) {
+    registry.add("spring.data.mongodb.uri", mongoContainer::getReplicaSetUrl);
   }
 }
